Changes to the email reg modal on the index.html
Changed how it worked to also make use of the regex Mayara coded, and made that its own .js file that is modular (in that it runs a method upon success)

Had to make some changes to the index.css to suit.

Also removed some comments that were leftover from debugging and such.